moved utils directory
